pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.130.0",
    "crewai-tools>=0.47.1",
    "elasticsearch>=8.18.1",
    "elasticsearch7>=7.9.1",
    "fastapi[standard]>=0.115.13",
    "google-genai>=1.21.1",
    "huggingface-hub>=0.33.0",
    "langchain>=0.3.25",
    "langchain-elasticsearch>=0.3.2",
    "langchain-google-genai>=2.1.5",
    "langchain-huggingface>=0.3.0",
    "langchain-text-splitters>=0.3.8",
    "langchain-unstructured>=0.1.5",
    "langgraph>=0.4.8",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.9.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=4.1.0",
    "unstructured>=0.17.2",
    "unstructured-client>=0.25.9",
    "uvicorn>=0.34.3",
]
[project.scripts]
main = "main:main"


[dependency-groups]
dev = [
    "prisma>=0.15.0",
]

